ref #491 Modified attribute in use case for versioning data Attribute "data_version" updated from int to float

+++ src/version_data/infrastructure/dvc_data_versioner.py
@@ -8,7 +8,7 @@ logger = logging.getLogger(__name__)
 
 
 class DVCDataVersioner():
-    def __init__(self, relative_data_path: str, data_name: str, data_version: int,
+    def __init__(self, relative_data_path: str, data_name: str, data_version: float,
                  git_remote_name: str, git_branch_name: str):
         self.relative_data_path = relative_data_path  # 'data/01_raw'
         self.data_name = data_name
